Move ResourceUrl to submodule
